Refactor: 셀렉션 생성 관련 쿼리 임포트 변경
