+++ src/hooks/useSelectionLocations.ts
@@ -1,6 +1,6 @@
 import { ISelectionLocation } from "@/models/selection.model";
 import { useEffect, useState } from "react";
-import { fetchSelectionLocations } from "./queries/selectionCreate.query";
+import { fetchSelectionLocations } from "@/http/selectionCreate.api";
 
 
 const useSelectionLocations = () => {
