+++ src/hooks/useSpotCategories.ts
@@ -1,6 +1,6 @@
 import { ISelectionSpotCategory } from "@/models/selection.model";
 import { useEffect, useState } from "react";
-import { fetchSelectionSpotCategories } from "./queries/selectionCreate.query";
+import { fetchSelectionSpotCategories } from "@/http/selectionCreate.api";
 
 
 const useSpotCategories = () => {
